+++ lib/alf/ext/relation.rb
@@ -4,3 +4,6 @@ DEE      = Alf::Relation::DEE
 def Relation(*args)
   Alf::Relation(*args)
 end
+def Tuple(*args)
+  Alf::Tuple(*args)
+end